Extract common scheduler code
The schedulers are probably well enough understood to extract common
code from them.  If this is not the case, we can always Inline Class
later.
